chore(localdebug): provide resolve local cert api in local env manager (#3651)
* chore(localdebug): provide resolve local cert api in local env manager

* chore: return isTrust equal to undefined when skip trust cert

* refactor: clean useless code
