+++ packages/fx-core/src/plugins/solution/fx-solution/debug/constants.ts
@@ -6,9 +6,3 @@ export class LaunchBrowser {
   public static readonly chrome: string = "pwa-chrome";
   public static readonly edge: string = "pwa-msedge";
 }
-
-export class LocalDebugCertificate {
-  public static readonly CertFileName: string = "localhost.crt";
-  public static readonly KeyFileName: string = "localhost.key";
-  public static readonly FriendlyName: string = "TeamsFx Development Certificate";
-}
