+++ packages/fx-core/src/plugins/solution/fx-solution/debug/provisionLocal.ts
@@ -34,7 +34,7 @@ import {
 } from "../../../../common/local/localEnvProvider";
 import { prepareLocalAuthService } from "./util/localService";
 import { getAllowedAppIds } from "../../../../common/tools";
-import { LocalCertificateManager } from "./util/certificate";
+import { LocalCertificateManager } from "../../../../common/local/localCertificateManager";
 
 export async function setupLocalDebugSettings(
   ctx: v2.Context,
@@ -235,7 +235,8 @@ export async function configLocalDebugSettings(
             localSettings.frontend.trustDevCert = trustDevCert;
           }
 
-          const certManager = new LocalCertificateManager(ctx, inputs);
+          const certManager = new LocalCertificateManager(ctx.userInteraction, ctx.logProvider);
+
           const localCert = await certManager.setupCertificate(trustDevCert);
           if (localCert) {
             localSettings.frontend.sslCertFile = localCert.certPath;
